requirements.txt
python-docx>=0.8.11
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
transformers>=4.40.0
sentencepiece>=0.2.0
indic-transliteration>=2.3.0
ctransformers>=0.2.26
faiss-cpu>=1.7.3
numpy>=1.23.0
scikit-learn>=1.2.0
